Remove nfs41_op_lock*.c
As with open, there is enough in common between NFSv4.0 and NFSv4.1
locking operations that we should not have two completely disjoint
functions to support them.

Eventually I'd like to get rid of --with-nfs4-minorversion=1 so we can
remove all the ifdefs that, essentially, duplicate the runtime version
check and double every piece of minor version specific code.