Make openflags a pointer in FSAL_open and FSAL_open_by_name
The FSAL requires the ability to inform the functions calling it of
the mode in which a file was opened.  In POSIX, closing one file
descriptor causes a process to lose all locks on that file regardless
of the descriptor holding the locks.

This could lead to a problem in which one client opens a file as
read-only and is granted a lock, and another client later opens the
file read-write.  Closing the file and reopening in read-write would
cause the first client's lock to be lost.  Opening a second file
descriptor would un-necessarily waste resources, if both clients
operated on the substrate filesystem with equivalent credentials
(i.e. if all file access were made as super-user.)

With this modification, the FSAL may open all files read-write and
indicate this fact to the calling function, avoiding the above
scenario.