Fix deadlock pointed out by Eitan Ben-Amos
Re-insert the get_located functionality for use in lookupp.  The
deadlock came about when a FSAL_lookup of ".." returned the same
handle as the entry being queried but the entry wasn't marked as a
root.  This would cause a lock order inversion in which
cache_inode_get would acquire the attribute lock while
cache_inode_lookupp held the content lock.