Implement caching AVL abstraction (avl_x).
Defines an abstraction closely modelled on the rbtree_x caching
rbtree adt from ntirpc (used in DRC).

The idea is to save code duplication, but with somewhat less
abstraction than HashTable or cih.    For example, the caller is
responsible for locking (but a choice of pthread lock primitives is
pre-defined), and there is no "latch."  Atomics are used for cache
accesses, so these can safely be performed under a shared lock.