Add shared Ceph mount indirection.
There is no defined mechanism for Ceph clients to interact with
multiple, distinct clusters yet (but presumably there will be).
For now, every object in a given Ceph env is by definition in the
same cluster, so sharing a single Ceph mount between exports is
the only case--but it should be the unmarked case when a way to
disambiguate Ceph clusters is added.  Then each export should have a
refcounted shared_ceph_mount object, and create_export (and later
destroy_export, etc) must do housekeeping accordingly.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>